fix:change id model change type int to double

+++ src/main/kotlin/server/web/casa/app/address/domain/model/City.kt
@@ -3,7 +3,7 @@ package server.web.casa.app.address.domain.model
 import server.web.casa.app.user.domain.model.User
 
 data class City(
-    val cityId : Int,
+    val cityId : Long,
     val country : Country,
     val name : String,
 //    val user: User
